refactor: RankFactory 생성
Rank enum 을 도메인 패키지에서 상위 패키지로 이동시키기 위해 진행
