+++ src/test/java/step2/domain/LotteryResultTest.java
@@ -2,6 +2,7 @@ package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.Rank;
 
 import static org.assertj.core.api.Assertions.assertThat;
